+++ backend/src/main/java/com/autoever/idle/domain/carType/repository/CarTypeRepository.java
@@ -1,4 +1,4 @@
-package com.autoever.idle.domain.carType;
+package com.autoever.idle.domain.carType.repository;
 
 import java.util.List;
