+++ backend/src/test/java/com/autoever/idle/domain/trim/TrimServiceTest.java
@@ -1,6 +1,6 @@
 package com.autoever.idle.domain.trim;
 
-import com.autoever.idle.domain.carType.CarTypeRepository;
+import com.autoever.idle.domain.carType.repository.CarTypeRepository;
 import com.autoever.idle.domain.category.functionCategory.repository.FunctionCategoryRepository;
 import com.autoever.idle.domain.category.functionCategory.dto.FunctionCategoryDto;
 import com.autoever.idle.domain.trim.dto.TrimDto;
